// src/extended/cores/types.ts
export interface ListResult<T, R = any> {
    /**
     * total searched count
     */
    total?: number;
    /**
     * max items count in the page
     */
    limit?: number;
    /**
     * current page number.
     */
    page?: number;
    /**
     * (optional) time took in sec
     */
    took?: number;
    /**
     * items searched
     */
    list: T[];

    /**
     * (optional) aggr list
     */
    aggr?: R | R[];
}

export interface AggrResult {
    [key: string]: { [key: string]: number };
}

export interface AggrKeyCount {
    /** name of key(or bucket name) */
    key: string;
    /** number of count */
    val: number;
}

/**
 * type `PaginatedListResult`
 */
export interface PaginatedListResult<T, R = string> extends ListResult<T, R> {
    /**
     * current page
     */
    page: number;
}

/**
 * type `ListParam`
 */
export interface ListParam {
    /**
     * max items count to be fetched
     */
    limit?: number;
    /**
     * (optional) sorting order
     *  - 'asc': older first
     *  - 'desc': newer first
     *  - string: extended sorting features
     */
    sort?: 'asc' | 'desc' | string;
}

/**
 * type `PaginateParam`
 */
export interface PaginateParam extends ListParam {
    /**
     * page # to fetch (0-indexed)
     */
    page?: number;
    /**
     * (optional) offset # from start.
     */
    offset?: number;
    /**
     * (optional) flag to filter by uid
     */
    uid?: string;
    /**
     * (optional) flag to filter by sid (only for admin role)
     */
    sid?: string;
}

/**
 * type `BulkUpdateBody`
 */
export interface BulkUpdateBody<T> extends BulkBody<T> {
    /**
     * list bulk update model with id
     */
    list: T[];
}

/**
 * body data for bulk
 */
export interface BulkBody<T> {
    list: T[];
}